1) Using RSSI values for sizing wavelets in this priority: – first I see if a RSSI value for tag was received in session — then accept this size – if no RSSI data for a specific tag, I'm using SIZE value from the CSV – if no RSSI nor CSV I'm using the size specified by UI
2) For RSSI sizing I've added the scale factor to UI – it can be useful in some cases
Currently fro RSSI sizing I'm using simple linear converting as: scale factor * (100 - RSSI) = size

Example:
With scale = 5 and RSSI = 80 the size will be: 5 * (100 - 80) = 100 pixels
For RSSI 60 it will be: 5 * (100 - 60) = 200 pixels
(60 is doubled of 80 as requested)

Later I can made a better formula to make sizes of wavelets closer to an average value
But need to be sure that we have corresponding sizes in result: RSSI / CSV / UI...
Maybe it should be simplified

P.S. I saw the RSSI values near the 90 too (89,90,91).

3) Now the color disc lifetime for TEMP_C events is increased and can be specified in UI (default is 15 seconds)
So, wavelet's rings (and value) will be disappeared within ~9 seconds as usually and color disc after 6 second.

4) I have added a check for preventing TEMP_C wavelets rendering replacing by other events: until a TEMP_C event is visible no other events for the same tag will not be rendered

5) RSSI events is not rendering (just using their values for sizing)